# go: render code to render path template