The client library should no longer depend on gax.PathTemplate.
Updates googleapis/gax-go#32, since gax.PathTemplate can then
be removed and there would be nothing left to verify.

In addition to rendering gax.PathTemplate is also capable
of parsing a rendered string and returning the template "arguments".
This feature was never exposed by toolkit clients.
If we need it later, we can still easily create parsing code,
like we generate rendering code in this commit.

The Java implementation of PathTemplate escapes URL when rendering.
The Go implementation does not, and so the functionality is not
included here.

While this probably won't matter in production,
benchmarking pubsub's PublisherTopicPath shows
```
benchmark                    old ns/op     new ns/op     delta
BenchmarkPathTemplate-12     1052          104           -90.11%
```